警告対応
(1) FileLength  型を明示的に 64 ビットで宣言

+++ Include/TsSplitter/Common/TsSplitterTypes.h
@@ -80,7 +80,7 @@ enum class  ErrCode
 **    ファイルの長さを表す型。
 **/
 
-typedef     size_t              FileLength;
+typedef     uint64_t            FileLength;
 
 //----------------------------------------------------------------
 /**
